+++ features/step_definitions/shared/xcode.rb
@@ -61,6 +61,16 @@ Then(/the scheme (.+) has a build setting (.+) with value (.+) for the configura
   assert(out.include?(search_for), "Couldn't find '#{search_for}'")
 end
 
+Then(/^the target (.+) should have the build phase (.+)$/) do |target_name, phase_name|
+  project = Xcodeproj::Project.open(@xcodeproj_path)
+  targets = project.targets
+  target = targets.detect { |t| t.name == target_name }
+  flunk("Target #{target_name} not found in the project") if target.nil?
+  build_phase = target.build_phases.detect { |b| b.display_name == phase_name }
+  flunk("The target #{target_name} doesn't have build phases") if build_phase.nil?
+  assert_equal phase_name, build_phase.name
+end
+
 Then(/^the target (.+) should have the build phase (.+) in the first position$/) do |target_name, phase_name|
   project = Xcodeproj::Project.open(@xcodeproj_path)
   targets = project.targets